Use long form clientid to persist clientid on storage for V4 recovery. Since clientid could be 1024 byte long and contains non printable characters, we devide the long form clientid into segments of 255 characters to create a directory sturcture to represent the clientid.
For example:
abcdefghijklmn -> abcd/efgh/ijkl/mn

+++ src/SAL/nfs4_recovery.c
@@ -153,26 +153,31 @@ int nfs_in_grace(void)
 void nfs4_create_clid_name(nfs_client_record_t *cl_rec,
 			   nfs_client_id_t *clientid, struct svc_req *svc)
 {
-	int i;
 	sockaddr_t sa;
 	char buf[SOCK_NAME_MAX + 1];
-	longlong_t cl_val = 0;
+	char cidstr[PATH_MAX];
+	struct display_buffer       dspbuf = {sizeof(cidstr), cidstr, cidstr};
+	int total_len;
 
-	clientid->cid_recov_dir = gsh_malloc(256);
-	if (clientid->cid_recov_dir == NULL) {
-		LogEvent(COMPONENT_CLIENTID, "Mem_Alloc FAILED");
-		return;
-	}
 	/* get the caller's IP addr */
 	if (copy_xprt_addr(&sa, svc->rq_xprt))
 		sprint_sockip(&sa, buf, SOCK_NAME_MAX);
 	else
 		strmaxcpy(buf, "Unknown", SOCK_NAME_MAX);
 
-	for (i = 0; i < cl_rec->cr_client_val_len; i++)
-		cl_val += cl_rec->cr_client_val[i];
-
-	snprintf(clientid->cid_recov_dir, 256, "%s-%llx", buf, cl_val);
+	if (convert_opaque_value_max_for_dir(&dspbuf, cl_rec->cr_client_val, cl_rec->cr_client_val_len, PATH_MAX) > 0) {
+		total_len = strlen(cidstr) + strlen(buf) + 2;
+		/* hold both long form clientid and IP */
+		clientid->cid_recov_dir = gsh_malloc(total_len);
+		if (clientid->cid_recov_dir == NULL) {
+			LogEvent(COMPONENT_CLIENTID, "Mem_Alloc FAILED");
+			return;
+		}
+		memset(clientid->cid_recov_dir, 0, total_len);
+		
+		(void) snprintf(clientid->cid_recov_dir, total_len, "%s-%s",
+				buf, cidstr);
+	}
 
 	LogDebug(COMPONENT_CLIENTID, "Created client name [%s]",
 		 clientid->cid_recov_dir);
@@ -191,11 +196,12 @@ void nfs4_create_clid_name(nfs_client_record_t *cl_rec,
 void nfs4_create_clid_name41(nfs_client_record_t *cl_rec,
 			     nfs_client_id_t *clientid)
 {
-	int i;
 	char buf[SOCK_NAME_MAX + 1];
-	longlong_t cl_val = 0;
+	char cidstr[PATH_MAX];
+	struct display_buffer       dspbuf = {sizeof(cidstr), cidstr, cidstr};
+	int total_len;
 
-	clientid->cid_recov_dir = gsh_malloc(256);
+	clientid->cid_recov_dir = gsh_malloc(PATH_MAX);
 	if (clientid->cid_recov_dir == NULL) {
 		LogEvent(COMPONENT_CLIENTID, "Mem_Alloc FAILED");
 		return;
@@ -203,10 +209,19 @@ void nfs4_create_clid_name41(nfs_client_record_t *cl_rec,
 	/* get the caller's IP addr */
 	sprint_sockip(&clientid->cid_client_addr, buf, sizeof(buf));
 
-	for (i = 0; i < cl_rec->cr_client_val_len; i++)
-		cl_val += cl_rec->cr_client_val[i];
+	if (convert_opaque_value_max_for_dir(&dspbuf, cl_rec->cr_client_val, cl_rec->cr_client_val_len, PATH_MAX) > 0) {
+		total_len = strlen(cidstr) + strlen(buf) + 2;
+		/* hold both long form clientid and IP */
+		clientid->cid_recov_dir = gsh_malloc(total_len);
+		if (clientid->cid_recov_dir == NULL) {
+			LogEvent(COMPONENT_CLIENTID, "Mem_Alloc FAILED");
+			return;
+		}
+		memset(clientid->cid_recov_dir, 0, total_len);
 
-	snprintf(clientid->cid_recov_dir, 256, "%s-%llx", buf, cl_val);
+		(void) snprintf(clientid->cid_recov_dir, total_len, "%s-%s",
+				buf, cidstr);
+	}
 
 	LogDebug(COMPONENT_CLIENTID, "Created client name [%s]",
 		 clientid->cid_recov_dir);
@@ -222,8 +237,9 @@ void nfs4_create_clid_name41(nfs_client_record_t *cl_rec,
  */
 void nfs4_add_clid(nfs_client_id_t *clientid)
 {
-	int err;
-	char path[PATH_MAX + 1];
+	int err = 0;
+	char path[PATH_MAX + 1] = {0}, segment[NAME_MAX + 1] = {0};
+	int length, position = 0;
 
 	if (clientid->cid_minorversion > 0)
 		nfs4_create_clid_name41(clientid->cid_client_record, clientid);
@@ -234,10 +250,32 @@ void nfs4_add_clid(nfs_client_id_t *clientid)
 		return;
 	}
 
-	snprintf(path, sizeof(path), "%s/%s", v4_recov_dir,
-		 clientid->cid_recov_dir);
+	/* break clientid down if it is greater than max dir name */
+	/* and create a directory hierachy to represent the clientid. */
+	snprintf(path, sizeof(path), "%s", v4_recov_dir);
 
-	err = mkdir(path, 0700);
+	length = strlen(clientid->cid_recov_dir);
+	while (position < length) {
+		/* if the (remaining) clientid is shorter than 255 */
+		/* create the last level of dir and break out */
+		int len = strlen(&clientid->cid_recov_dir[position]);
+		if (len <= NAME_MAX) {
+			strcat(path, "/");
+			strncat(path, &clientid->cid_recov_dir[position], len);
+			err = mkdir(path, 0700);
+			break;
+		}
+		/* if (remaining) clientid is longer than 255, */
+		/* get the next 255 bytes and create a subdir */
+		strncpy(segment, &clientid->cid_recov_dir[position], NAME_MAX);
+		strcat(path, "/");
+		strncat(path, segment, NAME_MAX);
+		err = mkdir(path, 0700);
+		if (err == -1 && errno != EEXIST)
+			break;
+		position += NAME_MAX;
+	}
+
 	if (err == -1 && errno != EEXIST) {
 		LogEvent(COMPONENT_CLIENTID,
 			 "Failed to create client in recovery dir (%s), errno=%d",
@@ -254,15 +292,52 @@ void nfs4_add_clid(nfs_client_id_t *clientid)
  *
  * @param[in] recov_dir Recovery directory
  */
-void nfs4_rm_clid(char *recov_dir)
+void nfs4_rm_clid(const char *recov_dir, char *parent_path, int position)
 {
 	int err;
-	char path[PATH_MAX + 1];
+	char *path;
+	char *segment;
+	int len, segment_len;
+	int total_len;
 
 	if (recov_dir == NULL)
 		return;
 
-	snprintf(path, sizeof(path), "%s/%s", v4_recov_dir, recov_dir);
+	len = strlen(recov_dir);
+	if (position == len)
+		return;
+
+	segment = gsh_malloc(NAME_MAX+1);
+	if (segment == NULL) {
+		LogEvent(COMPONENT_CLIENTID,
+			 "Failed to remove client in recovery dir (%s), ENOMEM",
+			  recov_dir);
+		return;
+	}
+
+	memset(segment, 0, NAME_MAX+1);
+	strncpy(segment, &recov_dir[position], NAME_MAX);
+	segment_len = strlen(segment);
+
+	/* allocate enough memory for the new part of the string */
+	/* which is parent path + '/' + new segment */
+	total_len = strlen(parent_path) + segment_len + 2;
+	path = gsh_malloc(total_len);
+	if (path == NULL) {
+		LogEvent(COMPONENT_CLIENTID,
+			 "Failed to remove client in recovery dir (%s), ENOMEM",
+			  recov_dir);
+		gsh_free(segment);
+		return;
+	}
+	memset(path, 0, total_len);
+	(void) snprintf(path, total_len, "%s/%s",
+			parent_path, segment);
+	/* free setment as it has no use now */
+	gsh_free(segment);
+
+	/* recursively remove the directory hirerchy which represent the clientid */
+	nfs4_rm_clid(recov_dir, path, position+segment_len);
 
 	err = rmdir(path);
 	if (err == -1) {
@@ -272,6 +347,7 @@ void nfs4_rm_clid(char *recov_dir)
 	} else {
 		LogDebug(COMPONENT_CLIENTID, "Removed client dir [%s]", path);
 	}
+	gsh_free(path);
 }
 
 /**
@@ -329,6 +405,16 @@ void nfs4_chk_clid(nfs_client_id_t *clientid)
 	pthread_mutex_unlock(&grace.g_mutex);
 }
 
+static void free_heap(char *path, char *new_path, char *build_clid)
+{
+	if (path)
+		gsh_free(path);
+	if (new_path)
+		gsh_free(new_path);
+	if (build_clid)
+		gsh_free(build_clid);
+}
+
 /**
  * @brief Create the client reclaim list
  *
@@ -346,47 +432,161 @@ void nfs4_chk_clid(nfs_client_id_t *clientid)
  *
  * @return POSIX error codes.
  */
-static int nfs4_read_recov_clids(DIR *dp, char *srcdir, bool takeover)
+static int nfs4_read_recov_clids(DIR *dp, const char *parent_path, char *clid_str, char *tgtdir, int takeover)
 {
 	struct dirent *dentp;
+	DIR *subdp;
 	clid_entry_t *new_ent;
-	char src[PATH_MAX + 1], dest[PATH_MAX + 1];
-	int rc;
+	char *path = NULL;
+	char *new_path = NULL;
+	char *build_clid = NULL;
+	int rc = 0;
+	int num = 0;
+	char *ptr, *ptr2;
+	char temp[10];
+	int cid_len, len;
+	int segment_len;
+	int total_len;
+	int total_tgt_len;
+	int total_clid_len;
 
 	dentp = readdir(dp);
 	while (dentp != NULL) {
 		/* don't add '.' and '..', or any '.*' entry */
 		if (dentp->d_name[0] != '.') {
-			new_ent = gsh_malloc(sizeof(clid_entry_t));
-			if (new_ent == NULL) {
-				LogEvent(COMPONENT_CLIENTID,
-					 "Unable to allocate memory.");
-				return -1;
+			num++;
+
+			/* construct the path by appending the subdir for the next */
+			/* readdir. This recursion keeps reading the subdirectory */
+			/* until reaching the end. */
+			segment_len = strlen(dentp->d_name);
+			total_len = segment_len + 2 + strlen(parent_path);
+			path = gsh_malloc(total_len);
+			/* if failed on this subdirectory, move to next */
+			/* we might be lucky */
+			if (path == NULL) {
+				LogEvent(COMPONENT_CLIENTID, "malloc faied errno=%d", errno);
+				continue;
 			}
-			strmaxcpy(new_ent->cl_name, dentp->d_name, 256);
-			glist_add(&grace.g_clid_list, &new_ent->cl_list);
-			LogDebug(COMPONENT_CLIENTID, "added %s to clid list",
-				 new_ent->cl_name);
-			if (srcdir != NULL) {
-				snprintf(src, sizeof(src), "%s/%s", srcdir,
-					 dentp->d_name);
-				snprintf(dest, sizeof(dest), "%s/%s",
-					 v4_old_dir, dentp->d_name);
-				if (takeover)
-					rc = mkdir(dest, 0700);
-				else
-					rc = rename(src, dest);
-				if (rc == -1) {
-					LogEvent(COMPONENT_CLIENTID,
-						 "Failed to make dir (%s), errno=%d",
-						 dest, errno);
+			memset(path, 0, total_len);
+
+			strcpy(path, parent_path);
+			strcat(path, "/");
+			strncat(path, dentp->d_name, segment_len);
+			/* if tgtdir is not NULL, we need to build nfs4old/currentnode */
+			if (tgtdir) {
+				total_tgt_len = segment_len + 2 + strlen(tgtdir);
+				new_path = gsh_malloc(total_tgt_len);
+				if (new_path == NULL) {
+					LogEvent(COMPONENT_CLIENTID, "malloc faied errno=%d", errno);
+					gsh_free(path);
+					continue;
+				}
+				memset(new_path, 0, total_tgt_len);
+				strcpy(new_path, tgtdir);
+				strcat(new_path, "/");
+				strncat(new_path, dentp->d_name, segment_len);
+				rc = mkdir(new_path, 0700);
+				if ((rc == -1) && (errno != EEXIST)) {
+					LogEvent(COMPONENT_CLIENTID, "mkdir %s faied errno=%d", new_path, errno);
 				}
 			}
+			/* keep building the clientid str by cursively */
+			/* reading the directory structure */
+			if (clid_str)
+				total_clid_len = segment_len + 1 + strlen(clid_str);
+			else
+				total_clid_len = segment_len + 1;
+			build_clid = gsh_malloc(total_clid_len);
+			if (build_clid == NULL) {
+				LogEvent(COMPONENT_CLIENTID, "malloc faied errno=%d", errno);
+				free_heap(path, new_path, NULL);
+				continue;
+			}
+			memset(build_clid, 0, total_clid_len);
+			if (clid_str)
+				strcpy(build_clid, clid_str);
+			strncat(build_clid, dentp->d_name, segment_len);
+			subdp = opendir(path);
+			if (subdp == NULL) {
+				LogEvent(COMPONENT_CLIENTID, "opendir %s failed errno=%d", dentp->d_name, errno);
+				free_heap(path, new_path, build_clid);
+				continue;
+			}
+
+			if (tgtdir)
+				rc = nfs4_read_recov_clids(subdp, path, build_clid, new_path, takeover);
+			else
+				rc = nfs4_read_recov_clids(subdp, path, build_clid, NULL, takeover);
+
+			if (new_path)
+				gsh_free(new_path);
+
+			/* after recursion, if the subdir has no non-hidden directory */
+			/* this is the end of this clientid str. Add the clientstr to the list. */
+			if (rc == 0) {
+				/* the clid format is <IP>-(clid-len:long-form-clid-in-string-formt)
+				 * make sure this reconstructed string is valid by comparing clid-len
+				 * and the actual long-form-clid length in the string.
+				 * This is to prevent getting incompleted strings that might exist
+				 * due to program crash.
+				 */
+				ptr = strchr(build_clid, '(');
+				if (ptr == NULL) {
+					LogEvent(COMPONENT_CLIENTID,
+							"invalid clid format: %s", build_clid);
+					free_heap(path, new_path, build_clid);
+					continue;
+				}
+				ptr2 = strchr(ptr, ':');
+				if (ptr2 == NULL) {
+					LogEvent(COMPONENT_CLIENTID,
+							"invalid clid format: %s", build_clid);
+					free_heap(path, new_path, build_clid);
+					continue;
+				}
+				len = ptr2-ptr-1;
+				if (len >= 9) {
+					LogEvent(COMPONENT_CLIENTID,
+							"invalid clid format: %s", build_clid);
+					free_heap(path, new_path, build_clid);
+					continue;
+				}
+				strncpy(temp, ptr+1, len);
+				temp[len] = 0;
+				cid_len = atoi(temp);
+				len = strlen(ptr2);
+				if ((len == (cid_len+2)) && (ptr2[len-1] == ')')) {
+					new_ent = gsh_malloc(sizeof(clid_entry_t));
+					if (new_ent == NULL) {
+						LogEvent(COMPONENT_CLIENTID,
+							"Unable to allocate memory.");
+						free_heap(path, new_path, build_clid);
+						continue;
+					}
+					strcpy(new_ent->cl_name, build_clid);
+					glist_add(&grace.g_clid_list, &new_ent->cl_list);
+					LogDebug(COMPONENT_CLIENTID, "added %s to clid list",
+						 new_ent->cl_name);
+				}
+			}
+			gsh_free(build_clid);
+			/* If this is not for takeover, remove the directory hierarchy */
+			/* that represent the current clientid */
+			if (!takeover) {
+				rc = rmdir(path);
+				if (rc == -1) {
+					LogEvent(COMPONENT_CLIENTID,
+						 "Failed to rmdir (%s), errno=%d",
+						 path, errno);
+				}
+			}
+			gsh_free(path);
 		}
 		dentp = readdir(dp);
 	}
 
-	return 0;
+	return num;
 }
 
 /**
@@ -422,7 +622,7 @@ static void nfs4_load_recov_clids_nolock(nfs_grace_start_t *gsp)
 				 v4_old_dir, errno);
 			return;
 		}
-		rc = nfs4_read_recov_clids(dp, NULL, 0);
+		rc = nfs4_read_recov_clids(dp, v4_old_dir, NULL, NULL, 0);
 		if (rc == -1) {
 			(void)closedir(dp);
 			LogEvent(COMPONENT_CLIENTID,
@@ -440,7 +640,7 @@ static void nfs4_load_recov_clids_nolock(nfs_grace_start_t *gsp)
 			return;
 		}
 
-		rc = nfs4_read_recov_clids(dp, v4_recov_dir, 0);
+		rc = nfs4_read_recov_clids(dp, v4_recov_dir, NULL, v4_old_dir, 0);
 		if (rc == -1) {
 			(void)closedir(dp);
 			LogEvent(COMPONENT_CLIENTID,
@@ -483,7 +683,7 @@ static void nfs4_load_recov_clids_nolock(nfs_grace_start_t *gsp)
 			return;
 		}
 
-		rc = nfs4_read_recov_clids(dp, path, 1);
+		rc = nfs4_read_recov_clids(dp, path, NULL, v4_old_dir, 1);
 		if (rc == -1) {
 			(void)closedir(dp);
 			LogEvent(COMPONENT_CLIENTID,
@@ -516,14 +716,16 @@ void nfs4_load_recov_clids(nfs_grace_start_t *gsp)
 /**
  * @brief Clean up recovery directory
  */
-void nfs4_clean_old_recov_dir(void)
+void nfs4_clean_old_recov_dir(char *parent_path)
 {
 	DIR *dp;
 	struct dirent *dentp;
-	char path[PATH_MAX + 1];
+	char *path = NULL;
 	int rc;
+	int segment_len;
+	int total_len;
 
-	dp = opendir(v4_old_dir);
+	dp = opendir(parent_path);
 	if (dp == NULL) {
 		LogEvent(COMPONENT_CLIENTID,
 			 "Failed to open old v4 recovery dir (%s), errno=%d",
@@ -536,14 +738,26 @@ void nfs4_clean_old_recov_dir(void)
 		if (dentp->d_name[0] == '.')
 			continue;
 
-		snprintf(path, sizeof(path), "%s/%s", v4_old_dir,
+		segment_len = strlen(dentp->d_name);
+		total_len = strlen(parent_path) + 2 + segment_len;
+		path = gsh_malloc(total_len);
+		if (path == NULL) {
+			LogEvent(COMPONENT_CLIENTID,
+				 "Unable to allocate memory.");
+			continue;
+		}
+		memset(path, 0, total_len);
+
+		snprintf(path, total_len, "%s/%s", parent_path,
 			 dentp->d_name);
 
+		nfs4_clean_old_recov_dir(path);
 		rc = rmdir(path);
 		if (rc == -1) {
 			LogEvent(COMPONENT_CLIENTID,
 				 "Failed to remove %s, errno=%d", path, errno);
 		}
+		gsh_free(path);
 	}
 	closedir(dp);
 }
